requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
openai-whisper>=20231117
edge-tts>=6.1.0
